+feature - Added filesupport when project is deleted
Project exist in two planes, a Data Plane and a directory plane. Each
project has directories that correspond with the project and should be
created, deleted, and renamed(updated) whenever aspects of the project
change. Since HTTP at it's core uses file directories for routing across
your domain, we are providing a mirror of that within the application